Tools for describing matrices

Tools for describing matrices

# S3 method for tbl_df
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

trunc_mat(x, n = NULL, width = NULL, n_extra = NULL)

Arguments

x
Object to show.

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

See also

tibble-package

Examples

trunc_mat(mtcars)

#> # data.frame [32 × 11]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> *  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   21.0     6 160.0   110  3.90 2.620 16.46     0     1     4     4
#> 2   21.0     6 160.0   110  3.90 2.875 17.02     0     1     4     4
#> 3   22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1
#> 4   21.4     6 258.0   110  3.08 3.215 19.44     1     0     3     1
#> 5   18.7     8 360.0   175  3.15 3.440 17.02     0     0     3     2
#> 6   18.1     6 225.0   105  2.76 3.460 20.22     1     0     3     1
#> 7   14.3     8 360.0   245  3.21 3.570 15.84     0     0     3     4
#> 8   24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2
#> 9   22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2
#> 10  19.2     6 167.6   123  3.92 3.440 18.30     1     0     4     4
#> # ... with 22 more rows

print(as_tibble(mtcars))

#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> *  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   21.0     6 160.0   110  3.90 2.620 16.46     0     1     4     4
#> 2   21.0     6 160.0   110  3.90 2.875 17.02     0     1     4     4
#> 3   22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1
#> 4   21.4     6 258.0   110  3.08 3.215 19.44     1     0     3     1
#> 5   18.7     8 360.0   175  3.15 3.440 17.02     0     0     3     2
#> 6   18.1     6 225.0   105  2.76 3.460 20.22     1     0     3     1
#> 7   14.3     8 360.0   245  3.21 3.570 15.84     0     0     3     4
#> 8   24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2
#> 9   22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2
#> 10  19.2     6 167.6   123  3.92 3.440 18.30     1     0     4     4
#> # ... with 22 more rows

print(as_tibble(mtcars), n = 1)

#> # A tibble: 32 × 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> * <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    21     6   160   110   3.9  2.62 16.46     0     1     4     4
#> # ... with 31 more rows

print(as_tibble(mtcars), n = 3)

#> # A tibble: 32 × 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> * <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  21.0     6   160   110  3.90 2.620 16.46     0     1     4     4
#> 2  21.0     6   160   110  3.90 2.875 17.02     0     1     4     4
#> 3  22.8     4   108    93  3.85 2.320 18.61     1     1     4     1
#> # ... with 29 more rows

print(as_tibble(mtcars), n = 100)

#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#> *  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   21.0     6 160.0   110  3.90 2.620 16.46     0     1     4     4
#> 2   21.0     6 160.0   110  3.90 2.875 17.02     0     1     4     4
#> 3   22.8     4 108.0    93  3.85 2.320 18.61     1     1     4     1
#> 4   21.4     6 258.0   110  3.08 3.215 19.44     1     0     3     1
#> 5   18.7     8 360.0   175  3.15 3.440 17.02     0     0     3     2
#> 6   18.1     6 225.0   105  2.76 3.460 20.22     1     0     3     1
#> 7   14.3     8 360.0   245  3.21 3.570 15.84     0     0     3     4
#> 8   24.4     4 146.7    62  3.69 3.190 20.00     1     0     4     2
#> 9   22.8     4 140.8    95  3.92 3.150 22.90     1     0     4     2
#> 10  19.2     6 167.6   123  3.92 3.440 18.30     1     0     4     4
#> 11  17.8     6 167.6   123  3.92 3.440 18.90     1     0     4     4
#> 12  16.4     8 275.8   180  3.07 4.070 17.40     0     0     3     3
#> 13  17.3     8 275.8   180  3.07 3.730 17.60     0     0     3     3
#> 14  15.2     8 275.8   180  3.07 3.780 18.00     0     0     3     3
#> 15  10.4     8 472.0   205  2.93 5.250 17.98     0     0     3     4
#> 16  10.4     8 460.0   215  3.00 5.424 17.82     0     0     3     4
#> 17  14.7     8 440.0   230  3.23 5.345 17.42     0     0     3     4
#> 18  32.4     4  78.7    66  4.08 2.200 19.47     1     1     4     1
#> 19  30.4     4  75.7    52  4.93 1.615 18.52     1     1     4     2
#> 20  33.9     4  71.1    65  4.22 1.835 19.90     1     1     4     1
#> 21  21.5     4 120.1    97  3.70 2.465 20.01     1     0     3     1
#> 22  15.5     8 318.0   150  2.76 3.520 16.87     0     0     3     2
#> 23  15.2     8 304.0   150  3.15 3.435 17.30     0     0     3     2
#> 24  13.3     8 350.0   245  3.73 3.840 15.41     0     0     3     4
#> 25  19.2     8 400.0   175  3.08 3.845 17.05     0     0     3     2
#> 26  27.3     4  79.0    66  4.08 1.935 18.90     1     1     4     1
#> 27  26.0     4 120.3    91  4.43 2.140 16.70     0     1     5     2
#> 28  30.4     4  95.1   113  3.77 1.513 16.90     1     1     5     2
#> 29  15.8     8 351.0   264  4.22 3.170 14.50     0     1     5     4
#> 30  19.7     6 145.0   175  3.62 2.770 15.50     0     1     5     6
#> 31  15.0     8 301.0   335  3.54 3.570 14.60     0     1     5     8
#> 32  21.4     4 121.0   109  4.11 2.780 18.60     1     1     4     2

if (!requireNamespace("nycflights13", quietly = TRUE))
  stop("Please install the nycflights13 package to run the rest of this example")

print(nycflights13::flights, n_extra = 2)

#> # A tibble: 336,776 × 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#> 1   2013     1     1      517            515         2      830            819
#> 2   2013     1     1      533            529         4      850            830
#> 3   2013     1     1      542            540         2      923            850
#> 4   2013     1     1      544            545        -1     1004           1022
#> 5   2013     1     1      554            600        -6      812            837
#> 6   2013     1     1      554            558        -4      740            728
#> 7   2013     1     1      555            600        -5      913            854
#> 8   2013     1     1      557            600        -3      709            723
#> 9   2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#> # ... with 336,766 more rows, and 11 more variables: arr_delay <dbl>,
#> #   carrier <chr>, ...

print(nycflights13::flights, width = Inf)

#> # A tibble: 336,776 × 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#> 1   2013     1     1      517            515         2      830            819
#> 2   2013     1     1      533            529         4      850            830
#> 3   2013     1     1      542            540         2      923            850
#> 4   2013     1     1      544            545        -1     1004           1022
#> 5   2013     1     1      554            600        -6      812            837
#> 6   2013     1     1      554            558        -4      740            728
#> 7   2013     1     1      555            600        -5      913            854
#> 8   2013     1     1      557            600        -3      709            723
#> 9   2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#>    arr_delay carrier flight tailnum origin  dest air_time distance  hour minute
#>        <dbl>   <chr>  <int>   <chr>  <chr> <chr>    <dbl>    <dbl> <dbl>  <dbl>
#> 1         11      UA   1545  N14228    EWR   IAH      227     1400     5     15
#> 2         20      UA   1714  N24211    LGA   IAH      227     1416     5     29
#> 3         33      AA   1141  N619AA    JFK   MIA      160     1089     5     40
#> 4        -18      B6    725  N804JB    JFK   BQN      183     1576     5     45
#> 5        -25      DL    461  N668DN    LGA   ATL      116      762     6      0
#> 6         12      UA   1696  N39463    EWR   ORD      150      719     5     58
#> 7         19      B6    507  N516JB    EWR   FLL      158     1065     6      0
#> 8        -14      EV   5708  N829AS    LGA   IAD       53      229     6      0
#> 9         -8      B6     79  N593JB    JFK   MCO      140      944     6      0
#> 10         8      AA    301  N3ALAA    LGA   ORD      138      733     6      0
#>              time_hour
#>                 <dttm>
#> 1  2013-01-01 05:00:00
#> 2  2013-01-01 05:00:00
#> 3  2013-01-01 05:00:00
#> 4  2013-01-01 05:00:00
#> 5  2013-01-01 06:00:00
#> 6  2013-01-01 05:00:00
#> 7  2013-01-01 06:00:00
#> 8  2013-01-01 06:00:00
#> 9  2013-01-01 06:00:00
#> 10 2013-01-01 06:00:00
#> # ... with 336,766 more rows