frame_data {tibble}	R Documentation

Row-wise tibble creation

Description

Create tibbles laying out the data in rows, rather than in columns. This is useful for small tables of data where readability is important. Please see tibble-package for a general introduction.

Usage

frame_data(...)

Arguments

...
Arguments specifying the structure of a tibble. Variable names should be formulas, and may only appear before the data.

Value

A tibble.

Examples

frame_data(
  ~colA, ~colB,
  "a",   1,
  "b",   2,
  "c",   3
)

# frame_data will create a list column if the value in each cell is
# not a scalar
frame_data(
  ~x,  ~y,
  "a", 1:3,
  "b", 4:6
)

[Package tibble version 1.1 Index]