frame_matrix {tibble}	R Documentation

Row-wise matrix creation

Description

Create matrices laying out the data in rows, similar to matrix(..., byrow = TRUE), with a nicer-to-read syntax. This is useful for small matrices, e.g. covariance matrices, where readability is important. The syntax is inspired by tribble().

Usage

frame_matrix(...)

Arguments

...
<dynamic-dots> Arguments specifying the structure of a frame_matrix. Column names should be formulas, and may only appear before the data. These arguments are processed with rlang::list2() and support unquote via !! and unquote-splice via !!!.

Value

A matrix.

See Also

See quasiquotation for more details on tidy dots semantics, i.e. exactly how the ... argument is processed.

Examples

frame_matrix(
  ~col1, ~col2,
  1,     3,
  5,     2
)

[Package tibble version 3.1.6 Index]