Get a glimpse of your data

This is like a transposed version of print(): columns run down the page, and data runs across. This makes it possible to see every column in a data frame. It's a little like str() applied to a data frame but it tries to show you as much data as possible. (And it always shows the underlying data, even when applied to a remote data source.)

glimpse(x, width = NULL, ...)

Arguments

x
An object to glimpse at.

width
Width of output: defaults to the setting of the option tibble.width (if finite) or the width of the console.

...
Other arguments passed on to individual methods.

Value

x original x is (invisibly) returned, allowing glimpse() to be used within a data pipe line.

S3 methods

glimpse is an S3 generic with a customised method for tbls and data.frames, and a default method that calls str().

Examples

glimpse(mtcars)

#> Observations: 32
#> Variables: 11
#> $ mpg  [3m[90m<dbl>[39m[23m 21.0, 21.0, 22.8, 21.4, 18.7, 18.1, 14.3, 24.4, 22.8, 19.2, 17.8…
#> $ cyl  [3m[90m<dbl>[39m[23m 6, 6, 4, 6, 8, 6, 8, 4, 4, 6, 6, 8, 8, 8, 8, 8, 8, 4, 4, 4, 4, 8…
#> $ disp [3m[90m<dbl>[39m[23m 160.0, 160.0, 108.0, 258.0, 360.0, 225.0, 360.0, 146.7, 140.8, 1…
#> $ hp   [3m[90m<dbl>[39m[23m 110, 110, 93, 110, 175, 105, 245, 62, 95, 123, 123, 180, 180, 18…
#> $ drat [3m[90m<dbl>[39m[23m 3.90, 3.90, 3.85, 3.08, 3.15, 2.76, 3.21, 3.69, 3.92, 3.92, 3.92…
#> $ wt   [3m[90m<dbl>[39m[23m 2.620, 2.875, 2.320, 3.215, 3.440, 3.460, 3.570, 3.190, 3.150, 3…
#> $ qsec [3m[90m<dbl>[39m[23m 16.46, 17.02, 18.61, 19.44, 17.02, 20.22, 15.84, 20.00, 22.90, 1…
#> $ vs   [3m[90m<dbl>[39m[23m 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0…
#> $ am   [3m[90m<dbl>[39m[23m 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0…
#> $ gear [3m[90m<dbl>[39m[23m 4, 4, 4, 3, 3, 3, 3, 4, 4, 4, 4, 3, 3, 3, 3, 3, 3, 4, 4, 4, 3, 3…
#> $ carb [3m[90m<dbl>[39m[23m 4, 4, 1, 1, 2, 1, 4, 2, 2, 4, 4, 3, 3, 3, 4, 4, 4, 1, 2, 1, 1, 2…

if (requireNamespace("nycflights13", quietly = TRUE)) {
  glimpse(nycflights13::flights)
}

#> Observations: 336,776
#> Variables: 19
#> $ year           [3m[90m<int>[39m[23m 2013, 2013, 2013, 2013, 2013, 2013, 2013, 2013, 2013, …
#> $ month          [3m[90m<int>[39m[23m 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ day            [3m[90m<int>[39m[23m 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ dep_time       [3m[90m<int>[39m[23m 517, 533, 542, 544, 554, 554, 555, 557, 557, 558, 558,…
#> $ sched_dep_time [3m[90m<int>[39m[23m 515, 529, 540, 545, 600, 558, 600, 600, 600, 600, 600,…
#> $ dep_delay      [3m[90m<dbl>[39m[23m 2, 4, 2, -1, -6, -4, -5, -3, -3, -2, -2, -2, -2, -2, -…
#> $ arr_time       [3m[90m<int>[39m[23m 830, 850, 923, 1004, 812, 740, 913, 709, 838, 753, 849…
#> $ sched_arr_time [3m[90m<int>[39m[23m 819, 830, 850, 1022, 837, 728, 854, 723, 846, 745, 851…
#> $ arr_delay      [3m[90m<dbl>[39m[23m 11, 20, 33, -18, -25, 12, 19, -14, -8, 8, -2, -3, 7, -…
#> $ carrier        [3m[90m<chr>[39m[23m "UA", "UA", "AA", "B6", "DL", "UA", "B6", "EV", "B6", …
#> $ flight         [3m[90m<int>[39m[23m 1545, 1714, 1141, 725, 461, 1696, 507, 5708, 79, 301, …
#> $ tailnum        [3m[90m<chr>[39m[23m "N14228", "N24211", "N619AA", "N804JB", "N668DN", "N39…
#> $ origin         [3m[90m<chr>[39m[23m "EWR", "LGA", "JFK", "JFK", "LGA", "EWR", "EWR", "LGA"…
#> $ dest           [3m[90m<chr>[39m[23m "IAH", "IAH", "MIA", "BQN", "ATL", "ORD", "FLL", "IAD"…
#> $ air_time       [3m[90m<dbl>[39m[23m 227, 227, 160, 183, 116, 150, 158, 53, 140, 138, 149, …
#> $ distance       [3m[90m<dbl>[39m[23m 1400, 1416, 1089, 1576, 762, 719, 1065, 229, 944, 733,…
#> $ hour           [3m[90m<dbl>[39m[23m 5, 5, 5, 5, 6, 5, 6, 6, 6, 6, 6, 6, 6, 6, 6, 5, 6, 6, …
#> $ minute         [3m[90m<dbl>[39m[23m 15, 29, 40, 45, 0, 58, 0, 0, 0, 0, 0, 0, 0, 0, 0, 59, …
#> $ time_hour      [3m[90m<dttm>[39m[23m 2013-01-01 05:00:00, 2013-01-01 05:00:00, 2013-01-01 …