tibble-options {tibble}	R Documentation

Package options

Description

Display options for tbl_df, used by trunc_mat() and (indirectly) by print.tbl().

Package options

tibble.print_max: Row number threshold: Maximum number of rows printed. Set to Inf to always print all rows. Default: 20.

tibble.print_min: Number of rows printed if row number threshold is exceeded. Default: 10.

tibble.width: Output width. Default: NULL (use width option).

tibble.max_extra_cols: Number of extra columns printed in reduced form. Default: 100.

pillar.bold: Use bold font, e.g. for column headers? This currently defaults to FALSE, because many terminal fonts have poor support for bold fonts.

pillar.subtle: Use subtle style, e.g. for insignificant digits? Default: TRUE.

pillar.neg: Highlight negative numbers? Default: TRUE.

pillar.sigfig: The number of significant digits that will be printed and highlighted, default: 3. Set the pillar.subtle option to FALSE to turn off highlighting of significant digits.

pillar.min_title_chars: The minimum number of characters for the column title, default: 15. Column titles may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of column titles.

[Package tibble version 1.4.2 Index]