view {tibble}	R Documentation

View an object

Description

Calls utils::View() on the input and returns it, invisibly. If the input is not a data frame, it is processed using a variant of as.data.frame(head(x, n)). A message is printed if the number of rows exceeds n. This function has no effect in noninteractive sessions.

Usage

view(x, title = NULL, ..., n = NULL)

Arguments

x
The object to display.

title
The title to use for the display, by default the deparsed expression is used.

...
Unused, must be empty.

n
Maximum number of rows to display. Only used if x is not a data frame. Uses the view_max option by default.

Details

The RStudio IDE overrides utils::View(), this is picked up correctly.

[Package tibble version 3.1.6 Index]