add_column {tibble}	R Documentation

Add columns to a data frame

Description

This is a convenient way to add one or more columns to an existing data frame.

Usage

add_column(.data, ..., .before = NULL, .after = NULL)

Arguments

.data
Data frame to append to.

...
Name-value pairs, passed on to tibble(). All values must have one element for each row in the data frame, or be of length 1. These arguments are passed on to tibble(), and therefore also support unquote via !! and unquote-splice via !!!.

.before, .after
One-based column index or column name where to add the new columns, default: after last column.

See Also

Other addition: add_row

Examples

# add_column ---------------------------------
df <- tibble(x = 1:3, y = 3:1)

add_column(df, z = -1:1, w = 0)

# You can't overwrite existing columns
## Not run: 
add_column(df, x = 4:6)

## End(Not run)
# You can't create new observations
## Not run: 
add_column(df, z = 1:5)

## End(Not run)

[Package tibble version 1.4.2 Index]