Coerce vectors, lists, and matrices to data frames

This is an S3 generic. Methods are included for:

data frames (adds "tbl_df" classes)

tibbles (returns unchanged input)

lists and vectors (treats each element as a column)

matrices

tables

default (coerces to a list)

as_tibble(x, ..., .rows = NULL, .name_repair = c("assert_unique",
  "unique", "syntactic", "none", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL))

# S3 method for data.frame
as_tibble(x, validate = TRUE, ..., .rows = NULL,
  .name_repair = c("assert_unique", "unique", "syntactic", "none",
  "minimal"), rownames = pkgconfig::get_config("tibble::rownames", NULL))

# S3 method for list
as_tibble(x, validate = TRUE, ..., .rows = NULL,
  .name_repair = c("assert_unique", "unique", "syntactic", "none",
  "minimal"))

# S3 method for matrix
as_tibble(x, ...)

# S3 method for table
as_tibble(x, `_n` = "n", ..., n = `_n`)

# S3 method for NULL
as_tibble(x, ...)

# S3 method for default
as_tibble(x, ...)

Arguments

x
A vector, list, or matrix. If a list, each element must have the same length.

...
Other arguments passed on to individual methods.

.rows
The number of rows, useful to create a 0-column tibble or just as an additional check.

.name_repair
Treatment of problematic column names:

"none": No name repair or checks, beyond basic existence,

"minimal": Same as "none",

"unique": Make sure names are unique and not empty,

"assert_unique": (default value), no name repair, but check they are unique,

"syntactic": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R).

See name-repair for more details on these terms and the strategies used to enforce them.

rownames
Treatment of existing row names (for data frames and matrices):

NULL: remove row names

NA: keep row names (default, for compatibility; will issue a warning if row names are present in the input).

A string: the name of the new column that will contain the existing row names, which are no longer present in the result.

_n, validate
For compatibility only, do not use for new code.

n
Name for count column, default: "n".

Details

Vectors and unknown input is coerced to a list (unlike as.data.frame()): for vectors, the returned tibble has as many columns as the vector has elements.

as.data.frame() is effectively a thin wrapper around data.frame, and hence is rather slow (because it calls data.frame() on each element before cbinding together). as_tibble is a new S3 generic with more efficient methods for matrices and data frames.

The default behavior is to silently remove row names. New code should explicitly convert row names to a new column using the rownames argument. For existing code that relies on row names remaining, call pkgconfig::set_config("tibble::rownames" = NA) in your script or in your package's .onLoad() function.

See also

Examples

l <- list(x = 1:500, y = runif(500), z = 500:1)
df <- as_tibble(l)

m <- matrix(rnorm(50), ncol = 5)
colnames(m) <- c("a", "b", "c", "d", "e")
df <- as_tibble(m)

as_tibble(1:3, .name_repair = "unique")

#> New names:
#> * `` -> ..1
#> * `` -> ..2
#> * `` -> ..3

#> # A tibble: 1 x 3
#>     ..1   ..2   ..3
#>   <int> <int> <int>
#> 1     1     2     3

# For list-like inputs, as_tibble() is considerably simpler than as.data.frame()

# NOT RUN {
if (requireNamespace("bench", quietly = TRUE)) {
  l2 <- replicate(26, sample(letters), simplify = FALSE)
  names(l2) <- letters
  bench::mark(
    as_tibble(l2, .name_repair = "syntactic"),
    as_tibble(l2, .name_repair = "unique"),
    as_tibble(l2, .name_repair = "none"),
    as_tibble(l2),
    as.data.frame(l2),
    check = FALSE
  )
}
# }