format.tbl {tibble}	R Documentation

Tools for describing matrices

Description

Tools for describing matrices

Usage

## S3 method for class 'tbl'
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

## S3 method for class 'tbl_df'
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

## S3 method for class 'tbl'
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

## S3 method for class 'tbl_df'
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

trunc_mat(x, n = NULL, width = NULL, n_extra = NULL)

Arguments

x
Object to show.

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

See Also

tibble-package

Examples

trunc_mat(mtcars)

print(as_tibble(mtcars))
print(as_tibble(mtcars), n = 1)
print(as_tibble(mtcars), n = 3)
print(as_tibble(mtcars), n = 100)

if (!requireNamespace("nycflights13", quietly = TRUE))
  stop("Please install the nycflights13 package to run the rest of this example")

print(nycflights13::flights, n_extra = 2)
print(nycflights13::flights, width = Inf)

[Package tibble version 1.4.2 Index]