Printing tibbles

One of the main features of the tbl_df class is the printing:

Tibbles only print as many rows and columns as fit on one screen, supplemented by a summary of the remaining rows and columns.

Tibble reveals the type of each column, which keeps the user informed about whether a variable is, e.g., <chr> or <fct> (character versus factor).

Printing can be tweaked for a one-off call by calling print() explicitly and setting arguments like n and width. More persistent control is available by setting the options described below.

# S3 method for tbl
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for tbl
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

trunc_mat(x, n = NULL, width = NULL, n_extra = NULL)

Arguments

x
Object to format or print.

...
Other arguments passed on to individual methods.

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

Package options

Options used by the tibble and pillar packages to format and print tbl_df objects. Used by the formatting workhorse trunc_mat() and, therefore, indirectly, by print.tbl().

tibble.print_max: Row number threshold: Maximum number of rows printed. Set to Inf to always print all rows. Default: 20.

tibble.print_min: Number of rows printed if row number threshold is exceeded. Default: 10.

tibble.width: Output width. Default: NULL (use width option).

tibble.max_extra_cols: Number of extra columns printed in reduced form. Default: 100.

pillar.bold: Use bold font, e.g. for column headers? This currently defaults to FALSE, because many terminal fonts have poor support for bold fonts.

pillar.subtle: Use subtle style, e.g. for row numbers and data types? Default: TRUE.

pillar.subtle_num: Use subtle style for insignificant digits? Default: FALSE, is also affected by the pillar.subtle option.

pillar.neg: Highlight negative numbers? Default: TRUE.

pillar.sigfig: The number of significant digits that will be printed and highlighted, default: 3. Set the pillar.subtle option to FALSE to turn off highlighting of significant digits.

pillar.min_title_chars: The minimum number of characters for the column title, default: 15. Column titles may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of column titles.

Examples

print(as_tibble(mtcars))

#> # A tibble: 32 x 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows

print(as_tibble(mtcars), n = 1)

#> # A tibble: 32 x 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    21     6   160   110   3.9  2.62  16.5     0     1     4     4
#> # … with 31 more rows

print(as_tibble(mtcars), n = 3)

#> # A tibble: 32 x 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  21       6   160   110  3.9   2.62  16.5     0     1     4     4
#> 2  21       6   160   110  3.9   2.88  17.0     0     1     4     4
#> 3  22.8     4   108    93  3.85  2.32  18.6     1     1     4     1
#> # … with 29 more rows

print(as_tibble(iris), n = 100)

#> # A tibble: 150 x 5
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>            <dbl>       <dbl>        <dbl>       <dbl> <fct>     
#>   1          5.1         3.5          1.4         0.2 setosa    
#>   2          4.9         3            1.4         0.2 setosa    
#>   3          4.7         3.2          1.3         0.2 setosa    
#>   4          4.6         3.1          1.5         0.2 setosa    
#>   5          5           3.6          1.4         0.2 setosa    
#>   6          5.4         3.9          1.7         0.4 setosa    
#>   7          4.6         3.4          1.4         0.3 setosa    
#>   8          5           3.4          1.5         0.2 setosa    
#>   9          4.4         2.9          1.4         0.2 setosa    
#>  10          4.9         3.1          1.5         0.1 setosa    
#>  11          5.4         3.7          1.5         0.2 setosa    
#>  12          4.8         3.4          1.6         0.2 setosa    
#>  13          4.8         3            1.4         0.1 setosa    
#>  14          4.3         3            1.1         0.1 setosa    
#>  15          5.8         4            1.2         0.2 setosa    
#>  16          5.7         4.4          1.5         0.4 setosa    
#>  17          5.4         3.9          1.3         0.4 setosa    
#>  18          5.1         3.5          1.4         0.3 setosa    
#>  19          5.7         3.8          1.7         0.3 setosa    
#>  20          5.1         3.8          1.5         0.3 setosa    
#>  21          5.4         3.4          1.7         0.2 setosa    
#>  22          5.1         3.7          1.5         0.4 setosa    
#>  23          4.6         3.6          1           0.2 setosa    
#>  24          5.1         3.3          1.7         0.5 setosa    
#>  25          4.8         3.4          1.9         0.2 setosa    
#>  26          5           3            1.6         0.2 setosa    
#>  27          5           3.4          1.6         0.4 setosa    
#>  28          5.2         3.5          1.5         0.2 setosa    
#>  29          5.2         3.4          1.4         0.2 setosa    
#>  30          4.7         3.2          1.6         0.2 setosa    
#>  31          4.8         3.1          1.6         0.2 setosa    
#>  32          5.4         3.4          1.5         0.4 setosa    
#>  33          5.2         4.1          1.5         0.1 setosa    
#>  34          5.5         4.2          1.4         0.2 setosa    
#>  35          4.9         3.1          1.5         0.2 setosa    
#>  36          5           3.2          1.2         0.2 setosa    
#>  37          5.5         3.5          1.3         0.2 setosa    
#>  38          4.9         3.6          1.4         0.1 setosa    
#>  39          4.4         3            1.3         0.2 setosa    
#>  40          5.1         3.4          1.5         0.2 setosa    
#>  41          5           3.5          1.3         0.3 setosa    
#>  42          4.5         2.3          1.3         0.3 setosa    
#>  43          4.4         3.2          1.3         0.2 setosa    
#>  44          5           3.5          1.6         0.6 setosa    
#>  45          5.1         3.8          1.9         0.4 setosa    
#>  46          4.8         3            1.4         0.3 setosa    
#>  47          5.1         3.8          1.6         0.2 setosa    
#>  48          4.6         3.2          1.4         0.2 setosa    
#>  49          5.3         3.7          1.5         0.2 setosa    
#>  50          5           3.3          1.4         0.2 setosa    
#>  51          7           3.2          4.7         1.4 versicolor
#>  52          6.4         3.2          4.5         1.5 versicolor
#>  53          6.9         3.1          4.9         1.5 versicolor
#>  54          5.5         2.3          4           1.3 versicolor
#>  55          6.5         2.8          4.6         1.5 versicolor
#>  56          5.7         2.8          4.5         1.3 versicolor
#>  57          6.3         3.3          4.7         1.6 versicolor
#>  58          4.9         2.4          3.3         1   versicolor
#>  59          6.6         2.9          4.6         1.3 versicolor
#>  60          5.2         2.7          3.9         1.4 versicolor
#>  61          5           2            3.5         1   versicolor
#>  62          5.9         3            4.2         1.5 versicolor
#>  63          6           2.2          4           1   versicolor
#>  64          6.1         2.9          4.7         1.4 versicolor
#>  65          5.6         2.9          3.6         1.3 versicolor
#>  66          6.7         3.1          4.4         1.4 versicolor
#>  67          5.6         3            4.5         1.5 versicolor
#>  68          5.8         2.7          4.1         1   versicolor
#>  69          6.2         2.2          4.5         1.5 versicolor
#>  70          5.6         2.5          3.9         1.1 versicolor
#>  71          5.9         3.2          4.8         1.8 versicolor
#>  72          6.1         2.8          4           1.3 versicolor
#>  73          6.3         2.5          4.9         1.5 versicolor
#>  74          6.1         2.8          4.7         1.2 versicolor
#>  75          6.4         2.9          4.3         1.3 versicolor
#>  76          6.6         3            4.4         1.4 versicolor
#>  77          6.8         2.8          4.8         1.4 versicolor
#>  78          6.7         3            5           1.7 versicolor
#>  79          6           2.9          4.5         1.5 versicolor
#>  80          5.7         2.6          3.5         1   versicolor
#>  81          5.5         2.4          3.8         1.1 versicolor
#>  82          5.5         2.4          3.7         1   versicolor
#>  83          5.8         2.7          3.9         1.2 versicolor
#>  84          6           2.7          5.1         1.6 versicolor
#>  85          5.4         3            4.5         1.5 versicolor
#>  86          6           3.4          4.5         1.6 versicolor
#>  87          6.7         3.1          4.7         1.5 versicolor
#>  88          6.3         2.3          4.4         1.3 versicolor
#>  89          5.6         3            4.1         1.3 versicolor
#>  90          5.5         2.5          4           1.3 versicolor
#>  91          5.5         2.6          4.4         1.2 versicolor
#>  92          6.1         3            4.6         1.4 versicolor
#>  93          5.8         2.6          4           1.2 versicolor
#>  94          5           2.3          3.3         1   versicolor
#>  95          5.6         2.7          4.2         1.3 versicolor
#>  96          5.7         3            4.2         1.2 versicolor
#>  97          5.7         2.9          4.2         1.3 versicolor
#>  98          6.2         2.9          4.3         1.3 versicolor
#>  99          5.1         2.5          3           1.1 versicolor
#> 100          5.7         2.8          4.1         1.3 versicolor
#> # … with 50 more rows

print(mtcars, width = 10)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

mtcars2 <- as_tibble(cbind(mtcars, mtcars), .name_repair = "unique")

#> New names:
#> * mpg -> mpg..1
#> * cyl -> cyl..2
#> * disp -> disp..3
#> * hp -> hp..4
#> * drat -> drat..5
#> * … and 17 more

print(mtcars2, n = 25, n_extra = 3)

#> # A tibble: 32 x 22
#>    mpg..1 cyl..2 disp..3 hp..4 drat..5 wt..6 qsec..7 vs..8 am..9 gear..10
#>     <dbl>  <dbl>   <dbl> <dbl>   <dbl> <dbl>   <dbl> <dbl> <dbl>    <dbl>
#>  1   21        6   160     110    3.9   2.62    16.5     0     1        4
#>  2   21        6   160     110    3.9   2.88    17.0     0     1        4
#>  3   22.8      4   108      93    3.85  2.32    18.6     1     1        4
#>  4   21.4      6   258     110    3.08  3.22    19.4     1     0        3
#>  5   18.7      8   360     175    3.15  3.44    17.0     0     0        3
#>  6   18.1      6   225     105    2.76  3.46    20.2     1     0        3
#>  7   14.3      8   360     245    3.21  3.57    15.8     0     0        3
#>  8   24.4      4   147.     62    3.69  3.19    20       1     0        4
#>  9   22.8      4   141.     95    3.92  3.15    22.9     1     0        4
#> 10   19.2      6   168.    123    3.92  3.44    18.3     1     0        4
#> 11   17.8      6   168.    123    3.92  3.44    18.9     1     0        4
#> 12   16.4      8   276.    180    3.07  4.07    17.4     0     0        3
#> 13   17.3      8   276.    180    3.07  3.73    17.6     0     0        3
#> 14   15.2      8   276.    180    3.07  3.78    18       0     0        3
#> 15   10.4      8   472     205    2.93  5.25    18.0     0     0        3
#> 16   10.4      8   460     215    3     5.42    17.8     0     0        3
#> 17   14.7      8   440     230    3.23  5.34    17.4     0     0        3
#> 18   32.4      4    78.7    66    4.08  2.2     19.5     1     1        4
#> 19   30.4      4    75.7    52    4.93  1.62    18.5     1     1        4
#> 20   33.9      4    71.1    65    4.22  1.84    19.9     1     1        4
#> 21   21.5      4   120.     97    3.7   2.46    20.0     1     0        3
#> 22   15.5      8   318     150    2.76  3.52    16.9     0     0        3
#> 23   15.2      8   304     150    3.15  3.44    17.3     0     0        3
#> 24   13.3      8   350     245    3.73  3.84    15.4     0     0        3
#> 25   19.2      8   400     175    3.08  3.84    17.0     0     0        3
#> # … with 7 more rows, and 12 more variables: carb..11 <dbl>, mpg..12 <dbl>,
#> #   cyl..13 <dbl>, …

trunc_mat(mtcars)

#> # Description: data.frame [32 × 11]
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>  * <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows

if (requireNamespace("nycflights13", quietly = TRUE)) {
  print(nycflights13::flights, n_extra = 2)
  print(nycflights13::flights, width = Inf)
}

#> # A tibble: 336,776 x 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#>  1  2013     1     1      517            515         2      830            819
#>  2  2013     1     1      533            529         4      850            830
#>  3  2013     1     1      542            540         2      923            850
#>  4  2013     1     1      544            545        -1     1004           1022
#>  5  2013     1     1      554            600        -6      812            837
#>  6  2013     1     1      554            558        -4      740            728
#>  7  2013     1     1      555            600        -5      913            854
#>  8  2013     1     1      557            600        -3      709            723
#>  9  2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#> # … with 336,766 more rows, and 11 more variables: arr_delay <dbl>,
#> #   carrier <chr>, …
#> # A tibble: 336,776 x 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#>  1  2013     1     1      517            515         2      830            819
#>  2  2013     1     1      533            529         4      850            830
#>  3  2013     1     1      542            540         2      923            850
#>  4  2013     1     1      544            545        -1     1004           1022
#>  5  2013     1     1      554            600        -6      812            837
#>  6  2013     1     1      554            558        -4      740            728
#>  7  2013     1     1      555            600        -5      913            854
#>  8  2013     1     1      557            600        -3      709            723
#>  9  2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#>    arr_delay carrier flight tailnum origin dest  air_time distance  hour minute
#>        <dbl> <chr>    <int> <chr>   <chr>  <chr>    <dbl>    <dbl> <dbl>  <dbl>
#>  1        11 UA        1545 N14228  EWR    IAH        227     1400     5     15
#>  2        20 UA        1714 N24211  LGA    IAH        227     1416     5     29
#>  3        33 AA        1141 N619AA  JFK    MIA        160     1089     5     40
#>  4       -18 B6         725 N804JB  JFK    BQN        183     1576     5     45
#>  5       -25 DL         461 N668DN  LGA    ATL        116      762     6      0
#>  6        12 UA        1696 N39463  EWR    ORD        150      719     5     58
#>  7        19 B6         507 N516JB  EWR    FLL        158     1065     6      0
#>  8       -14 EV        5708 N829AS  LGA    IAD         53      229     6      0
#>  9        -8 B6          79 N593JB  JFK    MCO        140      944     6      0
#> 10         8 AA         301 N3ALAA  LGA    ORD        138      733     6      0
#>    time_hour          
#>    <dttm>             
#>  1 2013-01-01 05:00:00
#>  2 2013-01-01 05:00:00
#>  3 2013-01-01 05:00:00
#>  4 2013-01-01 05:00:00
#>  5 2013-01-01 06:00:00
#>  6 2013-01-01 05:00:00
#>  7 2013-01-01 06:00:00
#>  8 2013-01-01 06:00:00
#>  9 2013-01-01 06:00:00
#> 10 2013-01-01 06:00:00
#> # … with 336,766 more rows