Printing tibbles

One of the main features of the tbl_df class is the printing:

Tibbles only print as many rows and columns as fit on one screen, supplemented by a summary of the remaining rows and columns.

Tibble reveals the type of each column, which keeps the user informed about whether a variable is, e.g., <chr> or <fct> (character versus factor).

Printing can be tweaked for a one-off call by calling print() explicitly and setting arguments like n and width. More persistent control is available by setting the options described below. See also vignette("digits", package = "pillar") for a comparison to base options, and num() and char() for creating columns with custom formatting options.

As of tibble 3.1.0, printing is handled entirely by the pillar package. If you implement a package that extend tibble, the printed output can be customized in various ways. See vignette("extending", package = "pillar") for details.

# S3 method for tbl_df
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for tbl_df
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

Arguments

x
Object to format or print.

...
Other arguments passed on to individual methods.

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns, this may be slow for very wide tibbles.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

Details

These options can be set via options() and queried via getOption(). For this, add a pillar. prefix (the package name and a dot) to the option name. Example: for an option foo, use options(pillar.foo = value) to set it and getOption("pillar.foo") to retrieve the current value. An option value of NULL means that the default is used.

Package options

The following options control printing of tbl and tbl_df objects:

tibble.print_max: Row number threshold: Maximum number of rows printed. Set to Inf to always print all rows. Default: 20.

tibble.print_min: Number of rows printed if row number threshold is exceeded. Default: 10.

tibble.width: Output width. Default: NULL (use width option).

tibble.max_extra_cols: Number of extra columns printed in reduced form. Default: 100.

The output uses color and highlighting according to the "cli.num_colors" option. Set it to 1 to suppress colored and highlighted output.

Options for the pillar package

print_max: Maximum number of rows printed, default: 20. Set to Inf to always print all rows. For compatibility reasons, getOption("tibble.print_max") and getOption("dplyr.print_max") are also consulted, this will be soft-deprecated in pillar v2.0.0.

print_min: Number of rows printed if the table has more than print_max rows, default: 10. For compatibility reasons, getOption("tibble.print_min") and getOption("dplyr.print_min") are also consulted, this will be soft-deprecated in pillar v2.0.0.

width: Output width. Default: NULL (use getOption("width")). This can be larger than getOption("width"), in this case the output of the table's body is distributed over multiple tiers for wide tibbles. For compatibility reasons, getOption("tibble.width") and getOption("dplyr.width") are also consulted, this will be soft-deprecated in pillar v2.0.0.

max_footer_lines: The maximum number of lines in the footer, default: 7. Set to Inf to turn off truncation of footer lines. The max_extra_cols option still limits the number of columns printed.

max_extra_cols: The maximum number of columns printed in the footer, default: 100. Set to Inf to show all columns. Set the more predictable max_footer_lines to control the number of footer lines instead.

bold: Use bold font, e.g. for column headers? This currently defaults to FALSE, because many terminal fonts have poor support for bold fonts.

subtle: Use subtle style, e.g. for row numbers and data types? Default: TRUE.

subtle_num: Use subtle style for insignificant digits? Default: FALSE, is also affected by the subtle option.

neg: Highlight negative numbers? Default: TRUE.

sigfig: The number of significant digits that will be printed and highlighted, default: 3. Set the subtle option to FALSE to turn off highlighting of significant digits.

min_title_chars: The minimum number of characters for the column title, default: 15. Column titles may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of column titles.

min_chars: The minimum number of characters wide to display character columns, default: 3. Character columns may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of character columns.

max_dec_width: The maximum allowed width for decimal notation, default: 13.

bidi: Set to TRUE for experimental support for bidirectional scripts. Default: FALSE. When this option is set, "left right override" and "first strong isolate" Unicode controls are inserted to ensure that text appears in its intended direction and that the column headings correspond to the correct columns.

Examples

print(as_tibble(mtcars))
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # … with 22 more rows
print(as_tibble(mtcars), n = 1)
#> # A tibble: 32 × 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    21     6   160   110   3.9  2.62  16.5     0     1     4     4
#> # … with 31 more rows
print(as_tibble(mtcars), n = 3)
#> # A tibble: 32 × 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  21       6   160   110  3.9   2.62  16.5     0     1     4     4
#> 2  21       6   160   110  3.9   2.88  17.0     0     1     4     4
#> 3  22.8     4   108    93  3.85  2.32  18.6     1     1     4     1
#> # … with 29 more rows

print(as_tibble(iris), n = 100)
#> # A tibble: 150 × 5
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width Species   
#>            <dbl>       <dbl>        <dbl>       <dbl> <fct>     
#>   1          5.1         3.5          1.4         0.2 setosa    
#>   2          4.9         3            1.4         0.2 setosa    
#>   3          4.7         3.2          1.3         0.2 setosa    
#>   4          4.6         3.1          1.5         0.2 setosa    
#>   5          5           3.6          1.4         0.2 setosa    
#>   6          5.4         3.9          1.7         0.4 setosa    
#>   7          4.6         3.4          1.4         0.3 setosa    
#>   8          5           3.4          1.5         0.2 setosa    
#>   9          4.4         2.9          1.4         0.2 setosa    
#>  10          4.9         3.1          1.5         0.1 setosa    
#>  11          5.4         3.7          1.5         0.2 setosa    
#>  12          4.8         3.4          1.6         0.2 setosa    
#>  13          4.8         3            1.4         0.1 setosa    
#>  14          4.3         3            1.1         0.1 setosa    
#>  15          5.8         4            1.2         0.2 setosa    
#>  16          5.7         4.4          1.5         0.4 setosa    
#>  17          5.4         3.9          1.3         0.4 setosa    
#>  18          5.1         3.5          1.4         0.3 setosa    
#>  19          5.7         3.8          1.7         0.3 setosa    
#>  20          5.1         3.8          1.5         0.3 setosa    
#>  21          5.4         3.4          1.7         0.2 setosa    
#>  22          5.1         3.7          1.5         0.4 setosa    
#>  23          4.6         3.6          1           0.2 setosa    
#>  24          5.1         3.3          1.7         0.5 setosa    
#>  25          4.8         3.4          1.9         0.2 setosa    
#>  26          5           3            1.6         0.2 setosa    
#>  27          5           3.4          1.6         0.4 setosa    
#>  28          5.2         3.5          1.5         0.2 setosa    
#>  29          5.2         3.4          1.4         0.2 setosa    
#>  30          4.7         3.2          1.6         0.2 setosa    
#>  31          4.8         3.1          1.6         0.2 setosa    
#>  32          5.4         3.4          1.5         0.4 setosa    
#>  33          5.2         4.1          1.5         0.1 setosa    
#>  34          5.5         4.2          1.4         0.2 setosa    
#>  35          4.9         3.1          1.5         0.2 setosa    
#>  36          5           3.2          1.2         0.2 setosa    
#>  37          5.5         3.5          1.3         0.2 setosa    
#>  38          4.9         3.6          1.4         0.1 setosa    
#>  39          4.4         3            1.3         0.2 setosa    
#>  40          5.1         3.4          1.5         0.2 setosa    
#>  41          5           3.5          1.3         0.3 setosa    
#>  42          4.5         2.3          1.3         0.3 setosa    
#>  43          4.4         3.2          1.3         0.2 setosa    
#>  44          5           3.5          1.6         0.6 setosa    
#>  45          5.1         3.8          1.9         0.4 setosa    
#>  46          4.8         3            1.4         0.3 setosa    
#>  47          5.1         3.8          1.6         0.2 setosa    
#>  48          4.6         3.2          1.4         0.2 setosa    
#>  49          5.3         3.7          1.5         0.2 setosa    
#>  50          5           3.3          1.4         0.2 setosa    
#>  51          7           3.2          4.7         1.4 versicolor
#>  52          6.4         3.2          4.5         1.5 versicolor
#>  53          6.9         3.1          4.9         1.5 versicolor
#>  54          5.5         2.3          4           1.3 versicolor
#>  55          6.5         2.8          4.6         1.5 versicolor
#>  56          5.7         2.8          4.5         1.3 versicolor
#>  57          6.3         3.3          4.7         1.6 versicolor
#>  58          4.9         2.4          3.3         1   versicolor
#>  59          6.6         2.9          4.6         1.3 versicolor
#>  60          5.2         2.7          3.9         1.4 versicolor
#>  61          5           2            3.5         1   versicolor
#>  62          5.9         3            4.2         1.5 versicolor
#>  63          6           2.2          4           1   versicolor
#>  64          6.1         2.9          4.7         1.4 versicolor
#>  65          5.6         2.9          3.6         1.3 versicolor
#>  66          6.7         3.1          4.4         1.4 versicolor
#>  67          5.6         3            4.5         1.5 versicolor
#>  68          5.8         2.7          4.1         1   versicolor
#>  69          6.2         2.2          4.5         1.5 versicolor
#>  70          5.6         2.5          3.9         1.1 versicolor
#>  71          5.9         3.2          4.8         1.8 versicolor
#>  72          6.1         2.8          4           1.3 versicolor
#>  73          6.3         2.5          4.9         1.5 versicolor
#>  74          6.1         2.8          4.7         1.2 versicolor
#>  75          6.4         2.9          4.3         1.3 versicolor
#>  76          6.6         3            4.4         1.4 versicolor
#>  77          6.8         2.8          4.8         1.4 versicolor
#>  78          6.7         3            5           1.7 versicolor
#>  79          6           2.9          4.5         1.5 versicolor
#>  80          5.7         2.6          3.5         1   versicolor
#>  81          5.5         2.4          3.8         1.1 versicolor
#>  82          5.5         2.4          3.7         1   versicolor
#>  83          5.8         2.7          3.9         1.2 versicolor
#>  84          6           2.7          5.1         1.6 versicolor
#>  85          5.4         3            4.5         1.5 versicolor
#>  86          6           3.4          4.5         1.6 versicolor
#>  87          6.7         3.1          4.7         1.5 versicolor
#>  88          6.3         2.3          4.4         1.3 versicolor
#>  89          5.6         3            4.1         1.3 versicolor
#>  90          5.5         2.5          4           1.3 versicolor
#>  91          5.5         2.6          4.4         1.2 versicolor
#>  92          6.1         3            4.6         1.4 versicolor
#>  93          5.8         2.6          4           1.2 versicolor
#>  94          5           2.3          3.3         1   versicolor
#>  95          5.6         2.7          4.2         1.3 versicolor
#>  96          5.7         3            4.2         1.2 versicolor
#>  97          5.7         2.9          4.2         1.3 versicolor
#>  98          6.2         2.9          4.3         1.3 versicolor
#>  99          5.1         2.5          3           1.1 versicolor
#> 100          5.7         2.8          4.1         1.3 versicolor
#> # … with 50 more rows

print(mtcars, width = 10)
#>                      mpg
#> Mazda RX4           21.0
#> Mazda RX4 Wag       21.0
#> Datsun 710          22.8
#> Hornet 4 Drive      21.4
#> Hornet Sportabout   18.7
#> Valiant             18.1
#> Duster 360          14.3
#> Merc 240D           24.4
#> Merc 230            22.8
#> Merc 280            19.2
#> Merc 280C           17.8
#> Merc 450SE          16.4
#> Merc 450SL          17.3
#> Merc 450SLC         15.2
#> Cadillac Fleetwood  10.4
#> Lincoln Continental 10.4
#> Chrysler Imperial   14.7
#> Fiat 128            32.4
#> Honda Civic         30.4
#> Toyota Corolla      33.9
#> Toyota Corona       21.5
#> Dodge Challenger    15.5
#> AMC Javelin         15.2
#> Camaro Z28          13.3
#> Pontiac Firebird    19.2
#> Fiat X1-9           27.3
#> Porsche 914-2       26.0
#> Lotus Europa        30.4
#> Ford Pantera L      15.8
#> Ferrari Dino        19.7
#> Maserati Bora       15.0
#> Volvo 142E          21.4
#>                     cyl
#> Mazda RX4             6
#> Mazda RX4 Wag         6
#> Datsun 710            4
#> Hornet 4 Drive        6
#> Hornet Sportabout     8
#> Valiant               6
#> Duster 360            8
#> Merc 240D             4
#> Merc 230              4
#> Merc 280              6
#> Merc 280C             6
#> Merc 450SE            8
#> Merc 450SL            8
#> Merc 450SLC           8
#> Cadillac Fleetwood    8
#> Lincoln Continental   8
#> Chrysler Imperial     8
#> Fiat 128              4
#> Honda Civic           4
#> Toyota Corolla        4
#> Toyota Corona         4
#> Dodge Challenger      8
#> AMC Javelin           8
#> Camaro Z28            8
#> Pontiac Firebird      8
#> Fiat X1-9             4
#> Porsche 914-2         4
#> Lotus Europa          4
#> Ford Pantera L        8
#> Ferrari Dino          6
#> Maserati Bora         8
#> Volvo 142E            4
#>                      disp
#> Mazda RX4           160.0
#> Mazda RX4 Wag       160.0
#> Datsun 710          108.0
#> Hornet 4 Drive      258.0
#> Hornet Sportabout   360.0
#> Valiant             225.0
#> Duster 360          360.0
#> Merc 240D           146.7
#> Merc 230            140.8
#> Merc 280            167.6
#> Merc 280C           167.6
#> Merc 450SE          275.8
#> Merc 450SL          275.8
#> Merc 450SLC         275.8
#> Cadillac Fleetwood  472.0
#> Lincoln Continental 460.0
#> Chrysler Imperial   440.0
#> Fiat 128             78.7
#> Honda Civic          75.7
#> Toyota Corolla       71.1
#> Toyota Corona       120.1
#> Dodge Challenger    318.0
#> AMC Javelin         304.0
#> Camaro Z28          350.0
#> Pontiac Firebird    400.0
#> Fiat X1-9            79.0
#> Porsche 914-2       120.3
#> Lotus Europa         95.1
#> Ford Pantera L      351.0
#> Ferrari Dino        145.0
#> Maserati Bora       301.0
#> Volvo 142E          121.0
#>                      hp
#> Mazda RX4           110
#> Mazda RX4 Wag       110
#> Datsun 710           93
#> Hornet 4 Drive      110
#> Hornet Sportabout   175
#> Valiant             105
#> Duster 360          245
#> Merc 240D            62
#> Merc 230             95
#> Merc 280            123
#> Merc 280C           123
#> Merc 450SE          180
#> Merc 450SL          180
#> Merc 450SLC         180
#> Cadillac Fleetwood  205
#> Lincoln Continental 215
#> Chrysler Imperial   230
#> Fiat 128             66
#> Honda Civic          52
#> Toyota Corolla       65
#> Toyota Corona        97
#> Dodge Challenger    150
#> AMC Javelin         150
#> Camaro Z28          245
#> Pontiac Firebird    175
#> Fiat X1-9            66
#> Porsche 914-2        91
#> Lotus Europa        113
#> Ford Pantera L      264
#> Ferrari Dino        175
#> Maserati Bora       335
#> Volvo 142E          109
#>                     drat
#> Mazda RX4           3.90
#> Mazda RX4 Wag       3.90
#> Datsun 710          3.85
#> Hornet 4 Drive      3.08
#> Hornet Sportabout   3.15
#> Valiant             2.76
#> Duster 360          3.21
#> Merc 240D           3.69
#> Merc 230            3.92
#> Merc 280            3.92
#> Merc 280C           3.92
#> Merc 450SE          3.07
#> Merc 450SL          3.07
#> Merc 450SLC         3.07
#> Cadillac Fleetwood  2.93
#> Lincoln Continental 3.00
#> Chrysler Imperial   3.23
#> Fiat 128            4.08
#> Honda Civic         4.93
#> Toyota Corolla      4.22
#> Toyota Corona       3.70
#> Dodge Challenger    2.76
#> AMC Javelin         3.15
#> Camaro Z28          3.73
#> Pontiac Firebird    3.08
#> Fiat X1-9           4.08
#> Porsche 914-2       4.43
#> Lotus Europa        3.77
#> Ford Pantera L      4.22
#> Ferrari Dino        3.62
#> Maserati Bora       3.54
#> Volvo 142E          4.11
#>                        wt
#> Mazda RX4           2.620
#> Mazda RX4 Wag       2.875
#> Datsun 710          2.320
#> Hornet 4 Drive      3.215
#> Hornet Sportabout   3.440
#> Valiant             3.460
#> Duster 360          3.570
#> Merc 240D           3.190
#> Merc 230            3.150
#> Merc 280            3.440
#> Merc 280C           3.440
#> Merc 450SE          4.070
#> Merc 450SL          3.730
#> Merc 450SLC         3.780
#> Cadillac Fleetwood  5.250
#> Lincoln Continental 5.424
#> Chrysler Imperial   5.345
#> Fiat 128            2.200
#> Honda Civic         1.615
#> Toyota Corolla      1.835
#> Toyota Corona       2.465
#> Dodge Challenger    3.520
#> AMC Javelin         3.435
#> Camaro Z28          3.840
#> Pontiac Firebird    3.845
#> Fiat X1-9           1.935
#> Porsche 914-2       2.140
#> Lotus Europa        1.513
#> Ford Pantera L      3.170
#> Ferrari Dino        2.770
#> Maserati Bora       3.570
#> Volvo 142E          2.780
#>                      qsec
#> Mazda RX4           16.46
#> Mazda RX4 Wag       17.02
#> Datsun 710          18.61
#> Hornet 4 Drive      19.44
#> Hornet Sportabout   17.02
#> Valiant             20.22
#> Duster 360          15.84
#> Merc 240D           20.00
#> Merc 230            22.90
#> Merc 280            18.30
#> Merc 280C           18.90
#> Merc 450SE          17.40
#> Merc 450SL          17.60
#> Merc 450SLC         18.00
#> Cadillac Fleetwood  17.98
#> Lincoln Continental 17.82
#> Chrysler Imperial   17.42
#> Fiat 128            19.47
#> Honda Civic         18.52
#> Toyota Corolla      19.90
#> Toyota Corona       20.01
#> Dodge Challenger    16.87
#> AMC Javelin         17.30
#> Camaro Z28          15.41
#> Pontiac Firebird    17.05
#> Fiat X1-9           18.90
#> Porsche 914-2       16.70
#> Lotus Europa        16.90
#> Ford Pantera L      14.50
#> Ferrari Dino        15.50
#> Maserati Bora       14.60
#> Volvo 142E          18.60
#>                     vs
#> Mazda RX4            0
#> Mazda RX4 Wag        0
#> Datsun 710           1
#> Hornet 4 Drive       1
#> Hornet Sportabout    0
#> Valiant              1
#> Duster 360           0
#> Merc 240D            1
#> Merc 230             1
#> Merc 280             1
#> Merc 280C            1
#> Merc 450SE           0
#> Merc 450SL           0
#> Merc 450SLC          0
#> Cadillac Fleetwood   0
#> Lincoln Continental  0
#> Chrysler Imperial    0
#> Fiat 128             1
#> Honda Civic          1
#> Toyota Corolla       1
#> Toyota Corona        1
#> Dodge Challenger     0
#> AMC Javelin          0
#> Camaro Z28           0
#> Pontiac Firebird     0
#> Fiat X1-9            1
#> Porsche 914-2        0
#> Lotus Europa         1
#> Ford Pantera L       0
#> Ferrari Dino         0
#> Maserati Bora        0
#> Volvo 142E           1
#>                     am
#> Mazda RX4            1
#> Mazda RX4 Wag        1
#> Datsun 710           1
#> Hornet 4 Drive       0
#> Hornet Sportabout    0
#> Valiant              0
#> Duster 360           0
#> Merc 240D            0
#> Merc 230             0
#> Merc 280             0
#> Merc 280C            0
#> Merc 450SE           0
#> Merc 450SL           0
#> Merc 450SLC          0
#> Cadillac Fleetwood   0
#> Lincoln Continental  0
#> Chrysler Imperial    0
#> Fiat 128             1
#> Honda Civic          1
#> Toyota Corolla       1
#> Toyota Corona        0
#> Dodge Challenger     0
#> AMC Javelin          0
#> Camaro Z28           0
#> Pontiac Firebird     0
#> Fiat X1-9            1
#> Porsche 914-2        1
#> Lotus Europa         1
#> Ford Pantera L       1
#> Ferrari Dino         1
#> Maserati Bora        1
#> Volvo 142E           1
#>                     gear
#> Mazda RX4              4
#> Mazda RX4 Wag          4
#> Datsun 710             4
#> Hornet 4 Drive         3
#> Hornet Sportabout      3
#> Valiant                3
#> Duster 360             3
#> Merc 240D              4
#> Merc 230               4
#> Merc 280               4
#> Merc 280C              4
#> Merc 450SE             3
#> Merc 450SL             3
#> Merc 450SLC            3
#> Cadillac Fleetwood     3
#> Lincoln Continental    3
#> Chrysler Imperial      3
#> Fiat 128               4
#> Honda Civic            4
#> Toyota Corolla         4
#> Toyota Corona          3
#> Dodge Challenger       3
#> AMC Javelin            3
#> Camaro Z28             3
#> Pontiac Firebird       3
#> Fiat X1-9              4
#> Porsche 914-2          5
#> Lotus Europa           5
#> Ford Pantera L         5
#> Ferrari Dino           5
#> Maserati Bora          5
#> Volvo 142E             4
#>                     carb
#> Mazda RX4              4
#> Mazda RX4 Wag          4
#> Datsun 710             1
#> Hornet 4 Drive         1
#> Hornet Sportabout      2
#> Valiant                1
#> Duster 360             4
#> Merc 240D              2
#> Merc 230               2
#> Merc 280               4
#> Merc 280C              4
#> Merc 450SE             3
#> Merc 450SL             3
#> Merc 450SLC            3
#> Cadillac Fleetwood     4
#> Lincoln Continental    4
#> Chrysler Imperial      4
#> Fiat 128               1
#> Honda Civic            2
#> Toyota Corolla         1
#> Toyota Corona          1
#> Dodge Challenger       2
#> AMC Javelin            2
#> Camaro Z28             4
#> Pontiac Firebird       2
#> Fiat X1-9              1
#> Porsche 914-2          2
#> Lotus Europa           2
#> Ford Pantera L         4
#> Ferrari Dino           6
#> Maserati Bora          8
#> Volvo 142E             2

mtcars2 <- as_tibble(cbind(mtcars, mtcars), .name_repair = "unique")
#> New names:
#> * mpg -> mpg...1
#> * cyl -> cyl...2
#> * disp -> disp...3
#> * hp -> hp...4
#> * drat -> drat...5
#> * ...
print(mtcars2, n = 25, n_extra = 3)
#> # A tibble: 32 × 22
#>    mpg...1 cyl...2 disp...3 hp...4 drat...5 wt...6 qsec...7 vs...8 am...9
#>      <dbl>   <dbl>    <dbl>  <dbl>    <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#>  1    21         6    160      110     3.9    2.62     16.5      0      1
#>  2    21         6    160      110     3.9    2.88     17.0      0      1
#>  3    22.8       4    108       93     3.85   2.32     18.6      1      1
#>  4    21.4       6    258      110     3.08   3.22     19.4      1      0
#>  5    18.7       8    360      175     3.15   3.44     17.0      0      0
#>  6    18.1       6    225      105     2.76   3.46     20.2      1      0
#>  7    14.3       8    360      245     3.21   3.57     15.8      0      0
#>  8    24.4       4    147.      62     3.69   3.19     20        1      0
#>  9    22.8       4    141.      95     3.92   3.15     22.9      1      0
#> 10    19.2       6    168.     123     3.92   3.44     18.3      1      0
#> 11    17.8       6    168.     123     3.92   3.44     18.9      1      0
#> 12    16.4       8    276.     180     3.07   4.07     17.4      0      0
#> 13    17.3       8    276.     180     3.07   3.73     17.6      0      0
#> 14    15.2       8    276.     180     3.07   3.78     18        0      0
#> 15    10.4       8    472      205     2.93   5.25     18.0      0      0
#> 16    10.4       8    460      215     3      5.42     17.8      0      0
#> 17    14.7       8    440      230     3.23   5.34     17.4      0      0
#> 18    32.4       4     78.7     66     4.08   2.2      19.5      1      1
#> 19    30.4       4     75.7     52     4.93   1.62     18.5      1      1
#> 20    33.9       4     71.1     65     4.22   1.84     19.9      1      1
#> 21    21.5       4    120.      97     3.7    2.46     20.0      1      0
#> 22    15.5       8    318      150     2.76   3.52     16.9      0      0
#> 23    15.2       8    304      150     3.15   3.44     17.3      0      0
#> 24    13.3       8    350      245     3.73   3.84     15.4      0      0
#> 25    19.2       8    400      175     3.08   3.84     17.0      0      0
#> # … with 7 more rows, and 13 more variables: gear...10 <dbl>, carb...11 <dbl>,
#> #   mpg...12 <dbl>, …

print(nycflights13::flights, n_extra = 2)
#> # A tibble: 336,776 × 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#>  1  2013     1     1      517            515         2      830            819
#>  2  2013     1     1      533            529         4      850            830
#>  3  2013     1     1      542            540         2      923            850
#>  4  2013     1     1      544            545        -1     1004           1022
#>  5  2013     1     1      554            600        -6      812            837
#>  6  2013     1     1      554            558        -4      740            728
#>  7  2013     1     1      555            600        -5      913            854
#>  8  2013     1     1      557            600        -3      709            723
#>  9  2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#> # … with 336,766 more rows, and 11 more variables: arr_delay <dbl>,
#> #   carrier <chr>, …
print(nycflights13::flights, width = Inf)
#> # A tibble: 336,776 × 19
#>     year month   day dep_time sched_dep_time dep_delay arr_time sched_arr_time
#>    <int> <int> <int>    <int>          <int>     <dbl>    <int>          <int>
#>  1  2013     1     1      517            515         2      830            819
#>  2  2013     1     1      533            529         4      850            830
#>  3  2013     1     1      542            540         2      923            850
#>  4  2013     1     1      544            545        -1     1004           1022
#>  5  2013     1     1      554            600        -6      812            837
#>  6  2013     1     1      554            558        -4      740            728
#>  7  2013     1     1      555            600        -5      913            854
#>  8  2013     1     1      557            600        -3      709            723
#>  9  2013     1     1      557            600        -3      838            846
#> 10  2013     1     1      558            600        -2      753            745
#>    arr_delay carrier flight tailnum origin dest  air_time distance  hour minute
#>        <dbl> <chr>    <int> <chr>   <chr>  <chr>    <dbl>    <dbl> <dbl>  <dbl>
#>  1        11 UA        1545 N14228  EWR    IAH        227     1400     5     15
#>  2        20 UA        1714 N24211  LGA    IAH        227     1416     5     29
#>  3        33 AA        1141 N619AA  JFK    MIA        160     1089     5     40
#>  4       -18 B6         725 N804JB  JFK    BQN        183     1576     5     45
#>  5       -25 DL         461 N668DN  LGA    ATL        116      762     6      0
#>  6        12 UA        1696 N39463  EWR    ORD        150      719     5     58
#>  7        19 B6         507 N516JB  EWR    FLL        158     1065     6      0
#>  8       -14 EV        5708 N829AS  LGA    IAD         53      229     6      0
#>  9        -8 B6          79 N593JB  JFK    MCO        140      944     6      0
#> 10         8 AA         301 N3ALAA  LGA    ORD        138      733     6      0
#>    time_hour          
#>    <dttm>             
#>  1 2013-01-01 05:00:00
#>  2 2013-01-01 05:00:00
#>  3 2013-01-01 05:00:00
#>  4 2013-01-01 05:00:00
#>  5 2013-01-01 06:00:00
#>  6 2013-01-01 05:00:00
#>  7 2013-01-01 06:00:00
#>  8 2013-01-01 06:00:00
#>  9 2013-01-01 06:00:00
#> 10 2013-01-01 06:00:00
#> # … with 336,766 more rows