Repair the names of a vector

tibble deals with a few levels of name repair:

minimal names exist. The names attribute is not NULL. The name of an unnamed element is "" and never NA.

unique names are minimal, have no duplicates, and are never empty (literally, no ""s).

syntactic names are unique and syntactic (see Details for more).

syntactic implies unique, unique implies minimal. These levels are nested.

The .name_repair argument of tibble() and as_tibble() refers to these levels. Alternatively, the user can pass their own name repair function. It should anticipate minimal names as input and should, likewise, return names that are at least minimal.

The "Life cycle" section explains the status of the existing functions tidy_names(), set_tidy_names(), and repair_names().

tidy_names(name, syntactic = FALSE, quiet = FALSE)

set_tidy_names(x, syntactic = FALSE, quiet = FALSE)

repair_names(x, prefix = "V", sep = "")

Arguments

name
A names attribute, usually a character vector.

syntactic
Should names be made syntactically valid via make.names()?

quiet
Whether to suppress messages about name repair.

x
A vector.

prefix
A string, the prefix to use for new column names.

sep
A string inserted between the column name and de-duplicating number.

Value

x with repaired names or a repaired version of name.

minimal names

minimal names exist. The names attribute is not NULL. The name of an unnamed element is "" and never NA.

tbl_df objects created by tibble() and as_tibble() have variable names that are minimal, at the very least, even when .name_repair = "none". Why? General name repair can be be implemented more simply if the baseline strategy ensures that names(x) returns a character vector of the correct length.

Examples:

Original names of a vector with length 3: NULL
                           minimal names: "" "" ""
    Original names: "x" NA
                           minimal names: "x" ""

Related: rlang::names2() returns the names of an object, after making them minimal.

unique names

unique names are minimal, have no duplicates, and are never empty (literally, no ""s).

You usually want unique variable names in a data.frame, because they ensure that any variable can be identified, uniquely, by its name. Indexing by name works.

There are many ways to make names unique. We append a suffix of the form ..j to any name that is "" or a duplicate, where j is the position. Why?

An absolute position j is more helpful than numbering within the columns that share a name. Context: troubleshooting data import with lots of columns and dysfunctional names.

We hypothesize that it's better have a "level playing field" when repairing names, i.e. if foo appears twice, they both get repaired, not just the second occurence.

Example:

Original names:    ""    "x"    "" "y"    "x"
  unique names: "..1" "x..2" "..3" "y" "x..5"

Why would you ever want minimal names, instead of unique or syntactic ones? Sometimes the first row of a data source -- allegedly variable names -- actually contains data and the resulting tibble will be reshaped with, e.g., tidyr::gather(). In this case, it is better to not munge the names at import.

Pre-existing suffixes of the form ..j are always stripped, prior to making names unique, i.e. reconstructing the suffixes. If this interacts poorly with your names, you should take control of name repair.

syntactic names

syntactic names are unique and syntactic (see make.names()), meaning they:

Have no duplicates (inherited from unique).

Consist of letters, numbers, and the dot . or underscore _ characters.

Start with a letter or the dot ., not followed by a number.

Not a reserved word.

syntactic names are easy to use "as is" in code. They do not require quoting and work well with nonstandard evaluation, such as list indexing via $ or in packages like dplyr and ggplot2.

There are many ways to make names syntactic. For example, we choose to define syntactic names as an extension of unique, i.e. syntactic implies unique. Why? Because the need for syntactic names is strongly associated with the need for uniqueness and this makes the name repair system simpler.

Life cycle

tidy_names(), set_tidy_names(), and repair_names() were early efforts to facilitate post hoc name repair in tibble, given that tibble() and as_tibble() did not do this.

From tibble v1.5.0, the .name_repair argument gives direct access to three specific levels of name repair: minimal, unique, and syntactic. We recommend that new code use this instead of tidy_names(), set_tidy_names(), or repair_names(). After a period of use, the repair stategies behind minimal, unique, and syntactic are likely to be exposed in standalone functions and this could affect the behaviour of tidy_names(). repair_names() should be considered deprecated.

tibble(..., `.name_repair = "unique"`)
## is preferred to
df <- tibble(...)
set_tidy_names(df, syntactic = FALSE)
    tibble(..., `.name_repair = "syntactic"`)
## is preferred to
df <- tibble(...)
set_tidy_names(df, syntactic = TRUE)

Examples

# NOT RUN {
## by default, duplicate names are not allowed
tibble(x = 1, x = 2)
# }
## you can authorize duplicate names
tibble(x = 1, x = 2, .name_repair = "none")

#> [90m# A tibble: 1 x 2[39m
#>       x     x
#>   [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m     1     2

## or request that the names be made unique
tibble(x = 1, x = 2, .name_repair = "unique")

#> New names:
#> * x -> x..1
#> * x -> x..2

#> [90m# A tibble: 1 x 2[39m
#>    x..1  x..2
#>   [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m     1     2

## by default, non-syntactic names are allowed
df <- tibble(`a 1` = 1, `a 2` = 2)
## because you can still index by name
df[["a 1"]]

#> [1] 1

df$`a 1`

#> [1] 1

## syntactic names are easier to work with, though, and you can request them
df <- tibble(`a 1` = 1, `a 2` = 2, .name_repair = "syntactic")

#> New names:
#> * `a 1` -> a.1
#> * `a 2` -> a.2

df$a.1

#> [1] 1

## you can specify your own name repair function
tibble(x = 1, x = 2, .name_repair = make.unique)

#> [90m# A tibble: 1 x 2[39m
#>       x   x.1
#>   [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m     1     2

fix_names <- function(x) gsub("%", " percent", x)
tibble(`25%` = 1, `75%` = 2, .name_repair = fix_names)

#> [90m# A tibble: 1 x 2[39m
#>   `25 percent` `75 percent`
#>          [3m[90m<dbl>[39m[23m        [3m[90m<dbl>[39m[23m
#> [90m1[39m            1            2

fix_names <- function(x) gsub("\\s+", "_", x)
tibble(`year 1` = 1, `year 2` = 2, .name_repair = fix_names)

#> [90m# A tibble: 1 x 2[39m
#>   year_1 year_2
#>    [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m1[39m      1      2

## the names attibute will be non-NULL, with "" as the default element
df <- as_tibble(list(1:3, letters[1:3]), .name_repair = "none")
names(df)

#> [1] "" ""