Tools for working with row names

While a tibble can have row names (e.g., when converting from a regular data frame), they are removed when subsetting with the [ operator. A warning will be raised when attempting to assign non-NULL row names to a tibble. Generally, it is best to avoid row names, because they are basically a character column with different semantics to every other column. These functions allow to you detect if a data frame has row names (has_rownames()), remove them (remove_rownames()), or convert them back-and-forth between an explicit column (rownames_to_column() and column_to_rownames()). Also included is rowid_to_column() which adds a column at the start of the dataframe of ascending sequential row ids starting at 1. Note that this will remove any existing row names.

has_rownames(df)

remove_rownames(df)

rownames_to_column(df, var = "rowname")

rowid_to_column(df, var = "rowid")

column_to_rownames(df, var = "rowname")

Arguments

df
A data frame

var
Name of column to use for rownames.

Details

In the printed output, the presence of row names is indicated by a star just above the row numbers.

Examples

has_rownames(mtcars)

#> [1] TRUE

has_rownames(iris)

#> [1] FALSE

has_rownames(remove_rownames(mtcars))

#> [1] FALSE

head(rownames_to_column(mtcars))

#>             rowname  mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1         Mazda RX4 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> 2     Mazda RX4 Wag 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> 3        Datsun 710 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> 4    Hornet 4 Drive 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> 5 Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> 6           Valiant 18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

mtcars_tbl <- as_tibble(rownames_to_column(mtcars))
mtcars_tbl

#> # A tibble: 32 x 12
#>    rowname      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <chr>      <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Mazda RX4   21.0  6.00   160 110    3.90  2.62  16.5  0     1.00  4.00  4.00
#>  2 Mazda RX4…  21.0  6.00   160 110    3.90  2.88  17.0  0     1.00  4.00  4.00
#>  3 Datsun 710  22.8  4.00   108  93.0  3.85  2.32  18.6  1.00  1.00  4.00  1.00
#>  4 Hornet 4 …  21.4  6.00   258 110    3.08  3.22  19.4  1.00  0     3.00  1.00
#>  5 Hornet Sp…  18.7  8.00   360 175    3.15  3.44  17.0  0     0     3.00  2.00
#>  6 Valiant     18.1  6.00   225 105    2.76  3.46  20.2  1.00  0     3.00  1.00
#>  7 Duster 360  14.3  8.00   360 245    3.21  3.57  15.8  0     0     3.00  4.00
#>  8 Merc 240D   24.4  4.00   147  62.0  3.69  3.19  20.0  1.00  0     4.00  2.00
#>  9 Merc 230    22.8  4.00   141  95.0  3.92  3.15  22.9  1.00  0     4.00  2.00
#> 10 Merc 280    19.2  6.00   168 123    3.92  3.44  18.3  1.00  0     4.00  4.00
#> # ... with 22 more rows

column_to_rownames(as.data.frame(mtcars_tbl))

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#>  [ reached getOption("max.print") -- omitted 23 rows ]