subsetting {tibble}	R Documentation

Subsetting tibbles

Description

Accessing columns, rows, or cells via $, [[, or [ is mostly similar to regular data frames. However, the behavior is different for tibbles and data frames in some cases:

[ always returns a tibble by default, even if only one column is accessed.

Partial matching of column names with $ and [[ is not supported, and NULL is returned. For $, a warning is given.

Only scalars (vectors of length one) or vectors with the same length as the number of rows can be used for assignment.

Rows outside of the tibble's boundaries cannot be accessed.

When updating with [[<- and [<-, type changes of entire columns are supported, but updating a part of a column requires that the new value is coercible to the existing type. See vec_slice() for the underlying implementation.

Unstable return type and implicit partial matching can lead to surprises and bugs that are hard to catch. If you rely on code that requires the original data frame behavior, coerce to a data frame via as.data.frame().

Usage

## S3 method for class 'tbl_df'
x$name

## S3 replacement method for class 'tbl_df'
x$name <- value

## S3 method for class 'tbl_df'
x[[i, j, ..., exact = TRUE]]

## S3 replacement method for class 'tbl_df'
x[[i, j, ...]] <- value

## S3 method for class 'tbl_df'
x[i, j, drop = FALSE, ...]

## S3 replacement method for class 'tbl_df'
x[i, j, ...] <- value

Arguments

x
A tibble.

name
A name or a string.

value
A value to store in a row, column, range or cell. Tibbles are stricter than data frames in what is accepted here.

i, j
Row and column indices. If j is omitted, i is used as column index.

...
Ignored.

exact
Ignored, with a warning.

drop
Coerce to a vector if fetching one column via tbl[, j] . Default FALSE, ignored when accessing a column via tbl[j] .

Details

For better compatibility with older code written for regular data frames, [ supports a drop argument which defaults to FALSE. New code should use [[ to turn a column into a vector.

Examples

df <- data.frame(a = 1:3, bc = 4:6)
tbl <- tibble(a = 1:3, bc = 4:6)

# Subsetting single columns:
df[, "a"]
tbl[, "a"]
tbl[, "a", drop = TRUE]
as.data.frame(tbl)[, "a"]

# Subsetting single rows with the drop argument:
df[1, , drop = TRUE]
tbl[1, , drop = TRUE]
as.list(tbl[1, ])

# Accessing non-existent columns:
df$b
tbl$b

df[["b", exact = FALSE]]
tbl[["b", exact = FALSE]]

df$bd <- c("n", "e", "w")
tbl$bd <- c("n", "e", "w")
df$b
tbl$b

df$b <- 7:9
tbl$b <- 7:9
df$b
tbl$b

# Identical behavior:
tbl[1, ]
tbl[1, c("bc", "a")]
tbl[, c("bc", "a")]
tbl[c("bc", "a")]
tbl["a"]
tbl$a
tbl[["a"]]

[Package tibble version 3.1.6 Index]