tibble-package {tibble}	R Documentation

tibble: Simple Data Frames

Description

Provides a 'tbl_df' class (the 'tibble') that provides stricter checking and better formatting than the traditional data frame.

Details

The S3 class tbl_df wraps a local data frame. The main advantage to using a tbl_df over a regular data frame is the printing: tbl objects only print a few rows and all the columns that fit on one screen, describing the rest of it as text.

Methods

tbl_df implements four important base methods:

print

By default only prints the first 10 rows (at most 20), and the columns that fit on screen; see print.tbl()

[

Does not simplify (drop) by default, returns a data frame

[[, $

Calls .subset2() directly, so is considerably faster. Returns NULL if column does not exist, $ warns.

Important functions

tibble() and tribble() for construction, as_tibble() for coercion, and print.tbl() and glimpse() for display.

Author(s)

Maintainer: Kirill Müller krlmlr+r@mailbox.org

Authors:

Hadley Wickham hadley@rstudio.com

Other contributors:

Romain Francois romain@r-enthusiasts.com [contributor]

RStudio [copyright holder]

See Also

Useful links:

http://tibble.tidyverse.org/

https://github.com/tidyverse/tibble

Report bugs at https://github.com/tidyverse/tibble/issues

Examples

tibble(a = 1:26, b = letters)
as_tibble(iris)

[Package tibble version 1.4.2 Index]