tibble: Simple Data Frames

Provides a 'tbl_df' class (the "tibble") that provides stricter checking and better formatting than the traditional data frame.

Details

The S3 class tbl_df wraps a local data frame. The main advantage to using a tbl_df over a regular data frame is the printing: tbl objects only print a few rows and all the columns that fit on one screen, describing the rest of it as text.

Methods

tbl_df implements four important base methods:

print

By default only prints the first 10 rows (at most 20), and the columns that fit on screen; see print.tbl_df()

[

Never simplifies (drops), so always returns data.frame

[[, $

Calls .subset2() directly, so is considerably faster. Returns NULL if column does not exist, $ warns.

Important functions

tibble() and tribble() for construction, as_tibble() for coercion, and print.tbl_df() and glimpse() for display.

Package options

Display options for tbl_df, used by trunc_mat() and (indirectly) by print.tbl_df().

tibble.print_max

Row number threshold: Maximum number of rows printed. Set to Inf to always print all rows. Default: 20.

tibble.print_min

Number of rows printed if row number threshold is exceeded. Default: 10.

tibble.width

Output width. Default: NULL (use width option).

tibble.max_extra_cols

Number of extra columns printed in reduced form. Default: 100.

See also

Useful links:

http://tibble.tidyverse.org/

https://github.com/tidyverse/tibble

Report bugs at https://github.com/tidyverse/tibble/issues