tibble: Simple Data Frames

Provides a 'tbl_df' class (the 'tibble') that provides stricter checking and better formatting than the traditional data frame.

Details

The tibble package provides utilities for handling tibbles, where "tibble" is a colloquial term for the S3 tbl_df class. The tbl_df class is a special case of the base data.frame. class, developed in response to lessons learned over many years of data analysis with data frames.

Tibble is the central data structure for the set of packages known as the tidyverse, including dplyr, ggplot2, tidyr, and readr.

General resources:

Website for the tibble package: https://tibble.tidyverse.org

Tibbles chapter in R for Data Science

Resources on specific topics:

Create a tibble: tibble(), as_tibble(), tribble(), enframe()

Inspect a tibble: print.tbl(), glimpse()

Details on the S3 tbl_df class: tbl_df

Package options

The following option is used for viewing tabular data with view():

tibble.view_max: Maximum number of rows shown if the input is not a data frame. Default: 1000.

The following options are used by the tibble and pillar packages to format and print tbl_df objects. Used by the formatting workhorse trunc_mat() and, therefore, indirectly, by print.tbl().

tibble.print_max: Row number threshold: Maximum number of rows printed. Set to Inf to always print all rows. Default: 20.

tibble.print_min: Number of rows printed if row number threshold is exceeded. Default: 10.

tibble.width: Output width. Default: NULL (use width option).

tibble.max_extra_cols: Number of extra columns printed in reduced form. Default: 100.

pillar.bold: Use bold font, e.g. for column headers? This currently defaults to FALSE, because many terminal fonts have poor support for bold fonts.

pillar.subtle: Use subtle style, e.g. for row numbers and data types? Default: TRUE.

pillar.subtle_num: Use subtle style for insignificant digits? Default: FALSE, is also affected by the pillar.subtle option.

pillar.neg: Highlight negative numbers? Default: TRUE.

pillar.sigfig: The number of significant digits that will be printed and highlighted, default: 3. Set the pillar.subtle option to FALSE to turn off highlighting of significant digits.

pillar.min_title_chars: The minimum number of characters for the column title, default: 15. Column titles may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of column titles.

See also

Author

Maintainer: Kirill Müller krlmlr+r@mailbox.org

Authors:

Hadley Wickham hadley@rstudio.com

Other contributors:

Romain Francois romain@r-enthusiasts.com [contributor]

Jennifer Bryan jenny@rstudio.com [contributor]

RStudio [copyright holder]